# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a lunch wallet server.",
        "title": "Lunch Wallet Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Lunch-wallet Team API Support",
            "url": "http://www.swagger.io/support",
            "email": "info@lunchwallet.com"
        },
        "license": {
            "name": "BSD",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/updatemeal/:id": {
            "put": {
                "description": "Update meal by clicking on the particular food. The id of each food is attached to the link/endpoint. When you click on the food to be updated, the id is gotten from the link/endpoint. It is an authorized route to only ADMIN",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Food"
                ],
                "summary": "Admin updates meal",
                "responses": {
                    "200": {
                        "description": "Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "not authorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}